rename Freight*Events to Carrier*Events
This comes from today's meeting with KN
