+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/CarrierEventCreator.java
@@ -21,50 +21,18 @@
 
 package org.matsim.contrib.freight.events;
 
-import java.util.Map;
-
 import org.matsim.api.core.v01.Id;
+import org.matsim.api.core.v01.events.Event;
+import org.matsim.api.core.v01.population.Activity;
 import org.matsim.contrib.freight.carrier.Carrier;
-import org.matsim.contrib.freight.carrier.CarrierService;
+import org.matsim.contrib.freight.carrier.ScheduledTour;
 import org.matsim.vehicles.Vehicle;
 
-import static org.matsim.contrib.freight.events.FreightEventAttributes.*;
+public interface CarrierEventCreator {
 
-/**
- * An event, that informs that a Freight {@link CarrierService} activity has ended.
- *
- * @author Tilman Matteis  - creating it for the use in Logistics / LogisticServiceProviders (LSP)s
- * @author Kai Martins-Turner (kturner) - integrating and adapting it into/for the MATSim freight contrib
- */
-public final class FreightServiceEndEvent extends AbstractFreightEvent{
-
-	public static final String EVENT_TYPE = "Freight service ends";
-	private final Id<CarrierService> serviceId;
-	private final double serviceDuration;
-
-	public FreightServiceEndEvent(double time, Id<Carrier> carrierId, CarrierService service, Id<Vehicle> vehicleId) {
-		super(time, carrierId, service.getLocationLinkId(), vehicleId);
-		this.serviceId = service.getId();
-		this.serviceDuration = service.getServiceDuration();
-	}
-
-	@Override public String getEventType() {
-		return EVENT_TYPE;
-	}
-
-	public Id<CarrierService> getServiceId() {
-		return serviceId;
-	}
-
-	public double getServiceDuration() {
-		return serviceDuration;
-	}
-
-	@Override
-	public Map<String, String> getAttributes() {
-		Map<String, String> attr = super.getAttributes();
-		attr.put(ATTRIBUTE_SERVICE_ID, serviceId.toString());
-		attr.put(ATTRIBUTE_SERVICE_DURATION, String.valueOf(serviceDuration));
-		return attr;
-	}
+	Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour, int activityCounter, Id<Vehicle> vehicleId);
+	// activityCounter is currently needed to get the correct "service" or "pickup" / "delivery" activity out auf the scheduled plan.
+	// It is well integrated in the {@link CarrierEventTracker}.
+	// Maybe it can be replaced by the correct freight-activity here --> move the getTourElement ... up
+	// kmt, Jun22
 }
