+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/CarrierEventCreatorUtils.java
@@ -21,27 +21,34 @@
 
 package org.matsim.contrib.freight.events;
 
-import org.matsim.api.core.v01.Id;
-import org.matsim.api.core.v01.events.ActivityStartEvent;
-import org.matsim.api.core.v01.events.Event;
-import org.matsim.api.core.v01.population.Activity;
-import org.matsim.contrib.freight.carrier.Carrier;
-import org.matsim.contrib.freight.carrier.FreightConstants;
-import org.matsim.contrib.freight.carrier.ScheduledTour;
-import org.matsim.contrib.freight.carrier.Tour.Pickup;
-import org.matsim.contrib.freight.carrier.Tour.TourElement;
-import org.matsim.vehicles.Vehicle;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 
-/*package-private*/  final class FreightShipmentPickupStartEventCreator implements FreightEventCreator {
+/**
+ * Utils for {@link CarrierEventCreator}s
+ *
+ * @author kturner
+ */
+public final class CarrierEventCreatorUtils {
 
-	@Override
-	public Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour, int activityCounter, Id<Vehicle> vehicleId) {
-		if(event instanceof ActivityStartEvent startEvent && FreightConstants.PICKUP.equals((startEvent).getActType()) ) {
-			TourElement element = scheduledTour.getTour().getTourElements().get(activityCounter);
-			if (element instanceof Pickup pickupActivity) {
-				return new FreightShipmentPickupStartEvent(event.getTime(), carrier.getId(), pickupActivity.getShipment(), vehicleId );
-			}
-		}
-		return null;
+	private CarrierEventCreatorUtils(){
 	}
+
+	/**
+	 * @return a collection of the standard freightEvent creators
+	 */
+	public static Collection<CarrierEventCreator> getStandardEventCreators(){
+		List<CarrierEventCreator> creators = new ArrayList<>();
+		creators.add(new CarrierServiceEndEventCreator());
+		creators.add(new CarrierServiceStartEventCreator());
+		creators.add(new CarrierShipmentDeliveryStartEventCreator());
+		creators.add(new CarrierShipmentDeliveryEndEventCreator());
+		creators.add(new CarrierShipmentPickupStartEventCreator());
+		creators.add(new CarrierShipmentPickupEndEventCreator());
+		creators.add(new CarrierTourEndEventCreator());
+		creators.add(new CarrierTourStartEventCreator());
+		return creators;
+	}
+
 }
