+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/CarrierServiceStartEventCreator.java
@@ -22,24 +22,24 @@
 package org.matsim.contrib.freight.events;
 
 import org.matsim.api.core.v01.Id;
-import org.matsim.api.core.v01.events.ActivityEndEvent;
+import org.matsim.api.core.v01.events.ActivityStartEvent;
 import org.matsim.api.core.v01.events.Event;
 import org.matsim.api.core.v01.population.Activity;
 import org.matsim.contrib.freight.carrier.Carrier;
 import org.matsim.contrib.freight.carrier.FreightConstants;
 import org.matsim.contrib.freight.carrier.ScheduledTour;
-import org.matsim.contrib.freight.carrier.Tour.Delivery;
+import org.matsim.contrib.freight.carrier.Tour.ServiceActivity;
 import org.matsim.contrib.freight.carrier.Tour.TourElement;
 import org.matsim.vehicles.Vehicle;
 
-/*package-private*/  final class FreightShipmentDeliveryEndEventCreator implements FreightEventCreator {
+/*package-private*/  final class CarrierServiceStartEventCreator implements CarrierEventCreator {
 
 	@Override
 	public Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour, int activityCounter, Id<Vehicle> vehicleId) {
-		if(event instanceof ActivityEndEvent endEvent && FreightConstants.DELIVERY.equals(endEvent.getActType()) ) {
+		if( event instanceof ActivityStartEvent startEvent && FreightConstants.SERVICE.equals(startEvent.getActType()) ){
 			TourElement element = scheduledTour.getTour().getTourElements().get(activityCounter);
-			if (element instanceof Delivery deliveryActivity) {
-				return new FreightShipmentDeliveryEndEvent(event.getTime(), carrier.getId(), deliveryActivity.getShipment(), vehicleId );
+			if( element instanceof ServiceActivity serviceActivity ) {
+				return new CarrierServiceStartEvent(event.getTime(), carrier.getId(), serviceActivity.getService(), vehicleId);
 			}
 		}
 		return null;
