+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/CarrierTourEndEventCreator.java
@@ -22,25 +22,21 @@
 package org.matsim.contrib.freight.events;
 
 import org.matsim.api.core.v01.Id;
-import org.matsim.api.core.v01.events.ActivityEndEvent;
+import org.matsim.api.core.v01.events.ActivityStartEvent;
 import org.matsim.api.core.v01.events.Event;
 import org.matsim.api.core.v01.population.Activity;
 import org.matsim.contrib.freight.carrier.Carrier;
 import org.matsim.contrib.freight.carrier.FreightConstants;
 import org.matsim.contrib.freight.carrier.ScheduledTour;
-import org.matsim.contrib.freight.carrier.Tour.ServiceActivity;
-import org.matsim.contrib.freight.carrier.Tour.TourElement;
 import org.matsim.vehicles.Vehicle;
 
-/*package-private*/  final class FreightServiceEndEventCreator implements FreightEventCreator {
+/*package-private*/ final class CarrierTourEndEventCreator implements CarrierEventCreator {
 
 	@Override
 	public Event createEvent(Event event, Carrier carrier, Activity activity, ScheduledTour scheduledTour, int activityCounter, Id<Vehicle> vehicleId) {
-		if(event instanceof ActivityEndEvent endEvent && FreightConstants.SERVICE.equals(endEvent.getActType())) {
-			TourElement element = scheduledTour.getTour().getTourElements().get(activityCounter);
-			if(element instanceof ServiceActivity serviceActivity) {
-				return new FreightServiceEndEvent(event.getTime(), carrier.getId(), serviceActivity.getService(), vehicleId);
-			}
+		if(event instanceof ActivityStartEvent startEvent && FreightConstants.END.equals(startEvent.getActType()) ) {
+				return new CarrierTourEndEvent(startEvent.getTime(), carrier.getId(), scheduledTour.getTour().getEndLinkId(), // TODO: If we have the tourId, we do not need to store the link here, kmt sep 22
+						vehicleId, scheduledTour.getTour().getId());
 		}
 		return null;
 	}
