+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/eventhandler/FreightShipmentPickupEventHandler.java
@@ -21,7 +21,7 @@
 
 package org.matsim.contrib.freight.events.eventhandler;
 
-import org.matsim.contrib.freight.events.FreightShipmentPickupEndEvent;
+import org.matsim.contrib.freight.events.CarrierShipmentPickupEndEvent;
 import org.matsim.core.events.handler.EventHandler;
 
 /**
@@ -31,5 +31,5 @@ import org.matsim.core.events.handler.EventHandler;
  *
  */
 public interface FreightShipmentPickupEventHandler extends EventHandler {
-	void handleEvent(FreightShipmentPickupEndEvent event);
+	void handleEvent(CarrierShipmentPickupEndEvent event);
 }
