+++ contribs/vsp/src/main/java/org/matsim/contrib/freight/analysis/CarrierLoadAnalysis.java
@@ -6,8 +6,8 @@ import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.Scenario;
 import org.matsim.api.core.v01.events.Event;
 import org.matsim.contrib.freight.carrier.Carriers;
-import org.matsim.contrib.freight.events.FreightShipmentDeliveryStartEvent;
-import org.matsim.contrib.freight.events.FreightShipmentPickupStartEvent;
+import org.matsim.contrib.freight.events.CarrierShipmentDeliveryStartEvent;
+import org.matsim.contrib.freight.events.CarrierShipmentPickupStartEvent;
 import org.matsim.core.events.handler.BasicEventHandler;
 import org.matsim.vehicles.Vehicle;
 import org.matsim.vehicles.VehicleType;
@@ -21,7 +21,7 @@ import java.util.LinkedHashMap;
 import java.util.LinkedList;
 import java.util.Map;
 
-import static org.matsim.contrib.freight.events.FreightEventAttributes.ATTRIBUTE_CAPACITYDEMAND;
+import static org.matsim.contrib.freight.events.CarrierEventAttributes.ATTRIBUTE_CAPACITYDEMAND;
 
 /**
  * @author Kai Martins-Turner (kturner)
@@ -39,9 +39,9 @@ public class CarrierLoadAnalysis implements BasicEventHandler {
 	}
 
 	@Override public void handleEvent(Event event) {
-		if (event.getEventType().equals(FreightShipmentPickupStartEvent.EVENT_TYPE)) {
+		if (event.getEventType().equals(CarrierShipmentPickupStartEvent.EVENT_TYPE)) {
 			handlePickup( event);
-		} if (event.getEventType().equals(FreightShipmentDeliveryStartEvent.EVENT_TYPE)) {
+		} if (event.getEventType().equals(CarrierShipmentDeliveryStartEvent.EVENT_TYPE)) {
 			handleDelivery(event);
 		}
 	}
