+++ contribs/vsp/src/main/java/org/matsim/contrib/freight/analysis/FreightAnalysisServiceTracking.java
@@ -24,8 +24,8 @@ import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.events.ActivityStartEvent;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.freight.carrier.*;
-import org.matsim.contrib.freight.events.FreightServiceEndEvent;
-import org.matsim.contrib.freight.events.FreightServiceStartEvent;
+import org.matsim.contrib.freight.events.CarrierServiceEndEvent;
+import org.matsim.contrib.freight.events.CarrierServiceStartEvent;
 
 import java.util.LinkedHashMap;
 
@@ -92,7 +92,7 @@ class FreightAnalysisServiceTracking {
 	}
 
 	// UNTESTED handling of LSP Service events that provided reliable info about driver and timestamps.
-	public void handleStartEvent(FreightServiceStartEvent event) {
+	public void handleStartEvent(CarrierServiceStartEvent event) {
 		if (carrierServiceTrackers.containsKey(event.getCarrierId())){
 			if (carrierServiceTrackers.get(event.getCarrierId()).serviceTrackers.containsKey(event.getServiceId())){
 				ServiceTracker service = carrierServiceTrackers.get(event.getCarrierId()).serviceTrackers.get(event.getServiceId());
@@ -101,7 +101,7 @@ class FreightAnalysisServiceTracking {
 			}
 		}
 	}
-	public void handleEndEvent(FreightServiceEndEvent event) {
+	public void handleEndEvent(CarrierServiceEndEvent event) {
 		if (carrierServiceTrackers.containsKey(event.getCarrierId())){
 			if (carrierServiceTrackers.get(event.getCarrierId()).serviceTrackers.containsKey(event.getServiceId())){
 				ServiceTracker service = carrierServiceTrackers.get(event.getCarrierId()).serviceTrackers.get(event.getServiceId());
