+++ contribs/vsp/src/main/java/org/matsim/contrib/freight/analysis/FreightAnalysisShipmentTracking.java
@@ -26,8 +26,8 @@ import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.freight.carrier.Carrier;
 import org.matsim.contrib.freight.carrier.CarrierShipment;
-import org.matsim.contrib.freight.events.FreightShipmentDeliveryEndEvent;
-import org.matsim.contrib.freight.events.FreightShipmentPickupEndEvent;
+import org.matsim.contrib.freight.events.CarrierShipmentDeliveryEndEvent;
+import org.matsim.contrib.freight.events.CarrierShipmentPickupEndEvent;
 
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
@@ -83,7 +83,7 @@ class FreightAnalysisShipmentTracking {
 		}
 	}
 // untested LSP Event handling for precise Shipment Tracking
-	public void trackPickedUpEvent(FreightShipmentPickupEndEvent event) {
+	public void trackPickedUpEvent(CarrierShipmentPickupEndEvent event) {
 		if (shipments.containsKey(event.getShipmentId())) {
 			shipments.get(event.getShipmentId()).pickUpTime = event.getTime();
 			//FixMe: Driver is no longer part of the events... kmt jul22
@@ -92,7 +92,7 @@ class FreightAnalysisShipmentTracking {
 	}
 
 
-	public void trackDeliveryEvent(FreightShipmentDeliveryEndEvent event) {
+	public void trackDeliveryEvent(CarrierShipmentDeliveryEndEvent event) {
 		if (shipments.containsKey(event.getShipmentId())){
 			ShipmentTracker shipmentTracker = shipments.get(event.getShipmentId());
 			shipmentTracker.deliveryTime=event.getTime();
